moved signin logic
